hopr_chain_rpc::rpc
Module safe_singleton

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

IsModuleEnabledCall

Container type for all input parameters for the isModuleEnabled function with signature isModuleEnabled(address) and selector 0x2d9ad53d

IsModuleEnabledReturn

Container type for all return fields from the isModuleEnabled function with signature isModuleEnabled(address) and selector 0x2d9ad53d

SafeSingleton

Statics

SAFESINGLETON_ABI

The parsed human-readable ABI of the contract.